+++ R/fab_dow.R
@@ -1,0 +1,88 @@
+#' Day of week figure(s)
+#'
+#' In a normal setting it may be that observations that occur at the weekend are
+#' indicative of data fabrication. \code{fab_dow} (short for fabrication, day of
+#' week), produces a plot that may help to identify problems. Customs vary in
+#' different countries, so that should be accounted for when interpreting these
+#' figures.
+#'
+#' @param data cdata frame containing \code{var} (and, optionally, \code{by}) variable(s)
+#' @param var string. Name of variable containing relevant dates or datetimes
+#'   (will be coerced to date via \code{as.Date})
+#' @param by  string. Name of variable denoting grouping
+#' @param dow_fmt format for day of week
+#' @param output output format \code{facet} combines figures via ggplot2::facet_wrap,
+#'   \code{list} returns a list of ggplot2 plots
+#' @param ... options passed to facet_wrap (see examples)
+#'
+#' @return list or ggplot2 object
+#' @importFrom purrr map
+#' @export
+#'
+#' @examples
+#' set.seed(234)
+#' dat <- data.frame(
+#'   x = Sys.Date() + sample(-20:19, 40, TRUE),
+#'   by = c(rep(1, 10), rep(2, 30))
+#' )
+#' dat %>% fab_dow("x")
+#' dat %>% fab_dow("x", "by")
+#' # free x scale
+#' dat %>% fab_dow("x", "by", scales = "free_x")
+#' #list of plots
+#' dat %>% fab_dow("x", "by", output = "list")
+#'
+
+fab_dow <- function(data, var, by = NULL, dow_fmt = "%a", output = c("facet", "list"), ...){
+
+  if(!is.character(var) | length(var) > 1) stop("'var' should be character of length 1")
+  if(!is.null(by) && (!is.character(by) | length(by) > 1)) stop("'by' should be character of length 1")
+  output <- match.arg(output)
+
+  tmp <- data %>%
+    mutate(x2 = as.Date(!!sym(var)))
+
+  if (sum(is.na(tmp$x2)) > sum(is.na(data[, var]))) warning("NAs introduced when coercing x to date")
+  if (all(is.na(tmp$x2))) stop("x should be coercible to date via as.Date(x)")
+
+  if(is.null(by)){
+    tmp <- tmp %>%
+      mutate(by = "Overall")
+  } else {
+    tmp <- tmp %>%
+      rename(by = !!sym(by))
+  }
+
+  dat <- tmp %>%
+    rename(date = !!sym(var)) %>%
+    mutate(dow = as.numeric(format(.data$date, format = "%u")),
+           dow = factor(dow, 7:1, rev(format(as.Date("2021-04-12") + 0:6, dow_fmt))))
+
+  fn <- function(dat){
+    dat %>%
+      filter(!is.na(dow)) %>%
+      ggplot(aes(x = dow)) +
+      geom_bar(show.legend = FALSE) +
+      scale_fill_discrete(drop = FALSE) +
+      scale_x_discrete(drop = FALSE) +
+      coord_flip() +
+      xlab("Day of week") +
+      ylab("Count")
+  }
+
+  if(length(unique(dat$by)) > 1){
+    if(output == "facet"){
+      out <- fn(dat) +
+        facet_wrap(~ by, ...)
+    }
+    if(output == "list"){
+      dat <- split(dat, dat$by)
+      out <- map(dat, fn)
+    }
+  } else {
+    out <- fn(dat)
+  }
+
+  return(out)
+}
+