+++ NEWS.md
@@ -1,3 +1,8 @@
+# kpitools 0.2.0
+
+* Addition of some tools for data fabrication (`fab_dow` and `fab_tod`)
+* Addition of a new vignette for data fabrication 
+
 # kpitools 0.1.1
 
 * Added colors assigned to risk level.